Add minimal dataset definition and analysis code
The purpose of this is just to check that a job request with out of
date codelists is failed in job-runner.

The codelist is deliberately out of date and represents an old download
of that codelist, without mapped-in VMPs.

The dataset definition and analysis script do very minimal things and
don't even use the out-of-date codelist (codelist checks don't take into
account whether the codelist is used in the study.